Updated chess engine; added Makefile and examples
Did more bug hunting within the chess engine. I'm fairly certain that
most of the major bugs plaguing the engine have been resolved. I still
need to include a check for stalemate, otherwise the game will go on
forever if each side only has a king. However, I'm not sure where to
include this. I may slightly alter the engine to determine checkmate and
stalemate at the end of each turn.

Wrote makefiles to compile the engine as a shared library. I initially
had an option to also compile this as a static library, but removed it
when it acted contrary to expectations.

Wrote some more examples, including a two-player game. There's a few
bugs in it, but it works well enough for bug hunting.