Refactored chessengine library
Contrary to my decision to rewrite the library in C, it was easier to
just switch to a better bitboard implementation. While this requires
more functionality and needs additional testing to verify its accuracy,
all of the functions appear to work without fault. I also added a
history module to make future perft testing easier.